pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Natalia Teplynina <nt@mnt9.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.6"
python-dotenv = "^0.20.0"
django-bootstrap4 = "^22.1"
psycopg2-binary = "^2.9.3"
gunicorn = "^20.1.0"
django-filter = "^22.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
